FEAT: 카드게임에 관련된 로직을 모두 LuckyGame 클래스에 통합
