+++ app/src/main/java/com/glacier/luckycardgamesofteer/MainActivity.kt
@@ -21,17 +21,21 @@ import com.glacier.luckycardgamesofteer.model.Participant
 class MainActivity : AppCompatActivity() {
 
     private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
-
+    private lateinit var luckyGame: LuckyGame
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
 
+        // 럭키 게임 객체 생성
+        luckyGame = LuckyGame()
         init()
 
     }
 
     fun init(){
+
+        luckyGame.initCard()
 
         // 참가자카드 리사이클러뷰 아이템 겹치게 보이기
         val deco = OverlapDecoration()
@@ -56,7 +60,8 @@ class MainActivity : AppCompatActivity() {
         binding.mbToggle.addOnButtonCheckedListener { _, checkedId, isChecked ->
             if(isChecked){
                 // 명수가 변경될 때 마다 카드를 새로 뽑는다
-                var cardList = initCard()
+                luckyGame.initCard()
+                val cardList = luckyGame.cardList
 
                 // 뽑은 카드를 한번 섞는다
                 cardList.shuffle()
@@ -101,107 +106,36 @@ class MainActivity : AppCompatActivity() {
         binding.mbToggle.check(R.id.btn_3people)
     }
 
-    fun initCard(): MutableList<Card> {
-        // 12개의 전체 카드 객체를 저장하기 위한 MutableList 선언
-        val cardList = mutableListOf<Card>()
-
-        // AnimalInfo enum중 랜덤하게 뽑은 카드를 총 12개 생성 후 리스트에 담는다. 숫자는 1~12로 지정
-        for(i in 1..12){
-            for (animal in Animal.values()) {
-                val pickedCard = Card(animal, i)
-                cardList.add(pickedCard)
-            }
-        }
-
-        // 생성한 카드 리스트를 리턴한다.
-        return cardList
-    }
-
-    fun shareCard(cardList_:MutableList<Card>, numOfParticipants: Int): MutableList<Participant>{
-        val participantList = mutableListOf<Participant>()
-
-        // 카드들을 수정할일이 생기기 때문에 (나누어 주는 과정) 깊은 복사를 하여 새로운 리스트로 만듦
-        var cardList = mutableListOf<Card>()
-        cardList.addAll(cardList_)
-
-        // 3명일때는 12번 카드는 제외시켜야함.
-        if(numOfParticipants == 3){
-            cardList = cardList.filter { it.num != 12 } as MutableList<Card>
-        }
-
-        // 참가자 한명당 카드를 배분함
-        for (i in 1..numOfParticipants) {
-            // 참가자 객체 생성
-            val participant = Participant(i.toString())
-
-            // 총 참가자 수에 따라 나누어 주는 카드 갯수가 다르기 때문에 분기
-            val numCardsPerParticipant = when (numOfParticipants) {
-                3 -> 8
-                4 -> 7
-                5 -> 6
-                else -> 0
-            }
-
-            // 카드 중에서 정해진 갯수만큼 들고온 다음, 뽑힌 카드는 삭제함
-            participant.addAllCard(cardList.take(numCardsPerParticipant))
-            cardList = cardList.drop(numCardsPerParticipant) as MutableList<Card>
-
-            // 참가자 생성 후 참가자 리스트에 추가함
-            participantList.add(participant)
-            //participant.showCards()
-        }
-
-        // 마찬가지로 남은 카드 갯수가 총 참가자 명수에 따라 다르기 때문에 분기
-        val numOfRemainCards = when (numOfParticipants) {
-            3 -> 9
-            4 -> 8
-            5 -> 6
-            else -> 0
-        }
-
-        // 남은 카드 따로 모으기
-        val remainCards = Participant("0")
-
-        // 마지막으로 남은 카드들을 저장
-        remainCards.addAllCard(cardList.take(numOfRemainCards))
-
-        // 남은 카드 또한 참가자 리스트에 넣고 최종 리턴
-        participantList.add(remainCards)
-        //remainCards.showCards()
-
-        return participantList
-    }
-
     fun setCardRecyclerView(cardList: MutableList<Card>, numOfParticipants: Int){
         // share cards
-        val participants = shareCard(cardList, numOfParticipants)
+        luckyGame.shareCard(numOfParticipants)
 
         // recyclerview 어댑터 설정
         binding.rv1.apply {
             layoutManager = LinearLayoutManager(applicationContext, RecyclerView.HORIZONTAL, false)
-            adapter = CardAdapter(participants[0].getCards(), true)
+            adapter = CardAdapter(luckyGame.participantList[0].getCards(), true)
         }
 
         binding.rv2.apply {
             layoutManager = LinearLayoutManager(applicationContext, RecyclerView.HORIZONTAL, false)
-            adapter = CardAdapter(participants[1].getCards(), false)
+            adapter = CardAdapter(luckyGame.participantList[1].getCards(), false)
         }
 
         binding.rv3.apply {
             layoutManager = LinearLayoutManager(applicationContext, RecyclerView.HORIZONTAL, false)
-            adapter = CardAdapter(participants[2].getCards(), false)
+            adapter = CardAdapter(luckyGame.participantList[2].getCards(), false)
         }
 
         if (numOfParticipants >= 4){
             binding.rv4.apply {
                 layoutManager = LinearLayoutManager(applicationContext, RecyclerView.HORIZONTAL, false)
-                adapter = CardAdapter(participants[3].getCards(), false)
+                adapter = CardAdapter(luckyGame.participantList[3].getCards(), false)
             }
         }
         if (numOfParticipants == 5){
             binding.rv5.apply {
                 layoutManager = LinearLayoutManager(applicationContext, RecyclerView.HORIZONTAL, false)
-                adapter = CardAdapter(participants[4].getCards(), false)
+                adapter = CardAdapter(luckyGame.participantList[4].getCards(), false)
             }
         }
 
@@ -212,7 +146,7 @@ class MainActivity : AppCompatActivity() {
                 4 -> layoutManager = GridLayoutManager(applicationContext, 4)
                 5 -> layoutManager = LinearLayoutManager(applicationContext, RecyclerView.HORIZONTAL, false)
             }
-            adapter = CardAdapter(participants[3].getCards(), false)
+            adapter = CardAdapter(luckyGame.participantList[3].getCards(), false)
         }
     }
