+++ app/src/test/java/com/glacier/luckycardgamesofteer/ExampleUnitTest.kt
@@ -12,6 +12,7 @@ import org.junit.Assert.*
 class ExampleUnitTest {
     @Test
     fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
+        assertEquals(8, 2 + 2)
     }
+
 }
